(ORCH-2296) Source shared gems from puppet-runtime
In preparation for the separation of pe-bolt-server project into its own
dedicated repo as part of ORCH-2296, this removes all configuration for
gems shared between the puppet-bolt and pe-bolt-server projects from
this repo and loads them from puppet-runtime instead.

- Gems used only by one of the projects are retained here for now.
- Subsequent PRs will remove pe-bolt-server and its dependencies from
  this repo and move them to a dedicated PE bolt-vanagon repo instead.

As part of these changes, the pe-bolt-server begins to use its own
runtime tarball.